web_app_with_auth/python/requirements.txt
Flask==2.0.2
python-dotenv
requests